requirements.txt
python-dateutil==2.8.2
google-cloud-storage==2.7.0
scikit-learn==1.2.0